add useTagToggle hook - unified hook interface to be consumed by TagSelector - so it doesn't need to worry about implementation details from several hooks for tag toggle logic
