+++ src/app.tsx
@@ -21,6 +21,7 @@ import { useUpdateChecker } from "./hooks/useUpdateChecker";
 import UpdateBanner from "./components/UpdateBanner";
 import { useMultiTrackTagging } from "./hooks/useMultiTrackTagging";
 import { useSmartPlaylists } from "./hooks/useSmartPlaylists";
+import { useTagToggle } from "./hooks/useTagToggle";
 
 function App() {
   const [showTagManager, setShowTagManager] = useState<boolean>(false);
@@ -30,7 +31,7 @@ function App() {
     tagData,
     lastSaved,
     isLoading,
-    toggleTagForTrack,
+    toggleTagSingleTrack,
     setRating,
     setEnergy,
     setBpm,
@@ -92,9 +93,11 @@ function App() {
     setMultiTagTracks,
     setLockedMultiTrackUri,
     setMultiTrackDraftTags,
-    toggleTag,
     cancelMultiTagging,
     selectedTagsForSelector,
+    toggleTagMultiTrack,
+    toggleStarRating,
+    toggleEnergyRating,
   } = useMultiTrackTagging();
 
   // Set up history tracking and URL param handling
@@ -106,6 +109,8 @@ function App() {
     setIsLocked,
     setLockedMultiTrackUri,
   });
+
+  const { handleToggleTag } = useTagToggle()
 
   const { updateInfo, dismissUpdate } = useUpdateChecker({
     currentVersion: packageJson.version,
@@ -140,23 +145,10 @@ function App() {
     };
   }, []);
 
+  // TODO: what is this
   const trackTags = isMultiTagging
     ? selectedTagsForSelector || []
     : tagData.tracks[activeTrack?.uri || ""]?.tags || [];
-
-  const handleToggleTag = (
-    categoryId: string,
-    subcategoryId: string,
-    tagId: string
-  ) => {
-    if (isMultiTagging) {
-      // Multi-track mode: Use the hook's logic
-      toggleTag(categoryId, subcategoryId, tagId);
-    } else if (activeTrack) {
-      // Single-track mode: Use the original logic directly
-      toggleTagForTrack(activeTrack.uri, categoryId, subcategoryId, tagId);
-    }
-  };
 
   const trackDataMap = useMemo(() => {
     return Object.fromEntries(
@@ -263,7 +255,7 @@ function App() {
                 onSetEnergy={(energy) => setEnergy(activeTrack.uri, energy)}
                 onSetBpm={(bpm) => setBpm(activeTrack.uri, bpm)}
                 onRemoveTag={(categoryId, subcategoryId, tagId) =>
-                  toggleTagForTrack(
+                  toggleTagSingleTrack(
                     activeTrack.uri,
                     categoryId,
                     subcategoryId,
